requirements.txt
PyVCF3~=1.0.3
tqdm~=4.66.4
aiohttp~=3.9.5
